refactor: keep REST handlers inline in runner.go
#24 had handleTrustBundleX509 and handleGetX509SVID inline in
runner.go. The delegated-API change had split them out into a new
internal/service/rest/ package; the body change (replacing the
"Hello: <stack>" stub with a real delegated-API-backed implementation)
didn't require that. Roll the package split back so this PR is a body
swap on the existing handler, not a re-architecture:

* runner.go regains the inline trustBundleCache and both handlers.
  Path-param is back to {stack} and lookup is cfg.Auth.LoadedStacks,
  matching #24.
* main.go drops the RESTDeps wiring; the handler reads
  cfg.Auth.LoadedStacks directly.
* internal/service/rest/ removed.
* Port-0 gRPC conditional from #28 was inadvertently lost during the
  cherry-pick; restored here.
